% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box_group.R
\name{create_dq_box_group}
\alias{create_dq_box_group}
\title{Create a dq box group}
\usage{
create_dq_box_group(session, ...)
}
\arguments{
\item{...}{box ids to be combined to one group}
}
\description{
Create a dq box group which automatically collapses all other
boxes whenever one box is opened.
}
\examples{
\donttest{library(shiny)
shinyApp(
  ui = fluidPage(
    init(),
    fluidRow(
      column(6,
        dq_box("Random Input1", dqSpace(250), "End of Content", id = "box1",
              title="Box1", width=12, collapsible=TRUE, openCallback=TRUE),
        dq_box("Random Input2", dqSpace(250), "End of Content", id = "box2",
              title="Box2", width=12, collapsible=TRUE, collapsed=TRUE, openCallback=TRUE),
        dq_box("Random Input3", dqSpace(250), "End of Content", id = "box3",
              title="Box3", width=12, collapsible=TRUE, collapsed=TRUE, openCallback=TRUE)
      )
    )
  ),
  server = function(input, output) {
    create_dq_box_group(session, "box1", "box2", "box3")
  }
)}
}
\author{
richard.kunze
}
