% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dq_drawer.R
\name{dq_drawer}
\alias{dq_drawer}
\title{Creates a drawer sidebar element}
\usage{
dq_drawer(id = NULL, ..., drawer_style = NULL, btn_style = NULL,
  direction = c("left", "right", "top", "bottom"))
}
\arguments{
\item{id}{optional element id, useful if current state is needed}

\item{...}{content elements, must be named since those will be used as button
labels}

\item{drawer_style, btn_style}{optional character specifying additional styles}

\item{direction}{optional, specifies the direction the drawer comes from, one
of c("left", "right", "top", "bottom)}
}
\value{
drawer element tag
}
\description{
Creates a drawer element with buttons to draw out different
content elements.
}
\examples{
## Only run examples in interactive R sessions
if (interactive()) {

library(shiny)
shinyApp(
  ui = fluidPage(
    dq_drawer(
      id = "myDrawer",
      Config = div("Some inputs to configurate things ..."),
      Red = div(style = "background: red;width: 100\%;height: 100\%;"),
      "Green Color!" = div("RED!!!!", style = "background: green;"),
      direction = "left"
    ),
    fluidRow(column(3, "Current page:", textOutput("drawerVal"), offset = 5))
  ),
  server = function(input, output) {
    output$drawerVal <- renderText(input$myDrawer)
  }
)

}
}
\author{
richard.kunze
}
