% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styling.R
\name{dq_footer}
\alias{dq_footer}
\title{Creates a custom footer}
\usage{
dq_footer(html = NULL, always_botton = TRUE)
}
\arguments{
\item{html}{HTML code to show within the footer}

\item{always_botton}{specify if this footer should always be on the bottom of the window, even if the body isn't as high as the window}
}
\value{
fluidRow with the custom footer
}
\description{
Creates a custom HTML footer with dq orange background and the given HTML code inside.
}
\examples{
\donttest{library(shiny)
shinyApp(
  ui = fluidPage(
    init(),
    dq_fullscreen_menu(),
    dq_header(),
    dq_footer()
  ),
  server = function(input, output) {
  }
)}
}
\author{
richard.kunze
}
