% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styling.R
\name{dq_fullscreen_menu}
\alias{dq_fullscreen_menu}
\title{Returns daqana fullscreen menu in pure HTML code}
\usage{
dq_fullscreen_menu(items)
}
\arguments{
\item{items}{list of items to show in the menu}
}
\value{
string representing the html code
}
\description{
Creates the daqana fullscreen menu as html code.
}
\examples{
\donttest{library(shiny)
shinyApp(
  ui = fluidPage(
    init(),
    dq_fullscreen_menu(),
    dq_header()
  ),
  server = function(input, output) {}
)}
}
\seealso{
Other html creating functions: \code{\link{dq_header}},
  \code{\link{dq_headline}}
}
\author{
richard.kunze
}
\concept{html creating functions}
