% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dq_handsontable.R
\name{dq_handsontable_output}
\alias{dq_handsontable_output}
\title{Adds an uiOutput}
\usage{
dq_handsontable_output(id, width = 12, offset = 0)
}
\arguments{
\item{id}{id of the element}

\item{width}{width of the table in bootstrap columns}

\item{offset}{optional offset of the column}
}
\value{
fluidrow containing a column with the output fields
}
\description{
Adds a fluidrow containing a column with the given width, ready to support a dq_handsontable_output
(or any other rendered ui).
}
\examples{
\donttest{library(shiny)
library(rhandsontable)
shinyApp(
  ui = fluidPage(
    dqshiny:::dq_handsontable_output("myCars")
  ),
  server = function(input, output) {
    dqshiny:::dq_render_handsontable("myCars", mtcars, filters = "T", reset = FALSE)
  }
)}
}
\author{
richard.kunze
}
