% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styling.R
\name{dq_header}
\alias{dq_header}
\title{Returns daqana header in pure HTML code}
\usage{
dq_header(logo = "dqshinyRes/img/logo_daqana.svg",
  link = "https://www.daqana.com", fullscreen_menu = TRUE)
}
\arguments{
\item{logo}{src of the daqana logo, or text}

\item{link}{hyperlink}

\item{fullscreen_menu}{add fullscreen menu to site}
}
\value{
string representing the html code
}
\description{
Creates the daqana header in dependency of the given log and title.
If no customer logo is given, just daqana logo will be in header.
}
\examples{
\donttest{library(shiny)
shinyApp(
  ui = fluidPage(
    init(),
    dq_header()
  ),
  server = function(input, output) {
  }
)}
}
\seealso{
Other html creating functions: \code{\link{dq_fullscreen_menu}},
  \code{\link{dq_headline}}
}
\author{
richard.kunze
}
