% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styling.R
\name{dq_headline}
\alias{dq_headline}
\title{Returns daqana headline}
\usage{
dq_headline(h1 = "", h4 = "")
}
\arguments{
\item{h1}{title, large text}

\item{h4}{description, small text}
}
\value{
shiny fluid row containing daqana headline
}
\description{
Creates a daqana headline
}
\examples{
\donttest{library(shiny)
shinyApp(
  ui = fluidPage(
    init(),
    dq_header(),
    dq_headline("My headline", "My Subheadline")
  ),
  server = function(input, output) {}
)}
}
\seealso{
Other html creating functions: \code{\link{dq_fullscreen_menu}},
  \code{\link{dq_header}}
}
\author{
david.breuer
}
\concept{html creating functions}
