% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dq_handsontable.R
\name{dq_render_handsontable}
\alias{dq_render_handsontable}
\title{Renders a dq handsontable}
\usage{
dq_render_handsontable(id, data, context, filters = "T", paged = TRUE,
  page_size = 25, reset = TRUE, table_params = list(readOnly = FALSE,
  stretchH = "all", contextMenu = FALSE), col_params = list(colWidths = 1,
  highlightCol = TRUE, highlightRow = TRUE, manualColumnResize = TRUE), ...)
}
\arguments{
\item{id}{id of the handsontable element, should be the same as the one used in the output element}

\item{data}{data to show in the table, should be a data.frame}

\item{context}{the context used to specify all ui elements used for this table}

\item{filters}{optional character vector, adds filters for each column, values must be one of c("T", "S", "R", "") to add a
TextFilter, SelectFilter, RangeFilter or none, vectors of length one will add a filter of this type for each column, set to
NULL to ommit filter row}

\item{paged}{optional logical, splits table into pages if TRUE}

\item{page_size}{optional numeric, number of items per page, can be one of 10, 25, 50, 100 or any other value
which will be added to this list}

\item{reset}{optional logical, specify whether to add a button to reset filters to initial values or not}

\item{table_params}{optional list, specify parameters to hand to rHandsontable table element}

\item{col_params}{optional list, specify parameters to hand to rHandsontable cols elements}

\item{...}{further optional lists to specify parameters to hand to rHandsontable col elements,
those lists have to be assigned to a name}
}
\value{
reactive values of the data in the table
}
\description{
Renders a handsontable into the given uiOutput id with the given data and parameters.
Can also contain several filters to filter the data and a feature to split the table into several pages with a given page size.
The function will also add all needed observeEvents to establish the required functionalities. If table is not readOnly, all
user inputs will automatically stored  and updated independent from any filters or pages.
}
\examples{
\donttest{library(shiny)
library(rhandsontable)
shinyApp(
  ui = fluidPage(
    init(),
    dq_handsontable_output("randomTable", 9)
  ),
  server = function(input, output, session) {
    hw <- c("Hello", "my", "funny", "world!")
    data <- data.frame(A=rep(hw, 500), B=rep(hw[c(2,3,4,1)], 500),
                       C=1:500, D=Sys.Date()-0:499, stringsAsFactors = F)
    dq_render_handsontable("randomTable", data, "rand", filters = c("S", "T", "R", "R"),
      pCol1=list(col=1, type = "dropdown", source = letters),
      pCol2=list(col=2:4, type = "dropdown", source = LETTERS))
  }
)}
}
\seealso{
\code{\link[rhandsontable:rhandsontable]{rhandsontable}}, \code{\link[rhandsontable:hot_cols]{hot_cols}} and \code{\link[rhandsontable:hot_col]{hot_col}}
}
\author{
richard.kunze
}
