% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dqshiny.R
\docType{package}
\name{dqshiny}
\alias{dqshiny}
\alias{dqshiny-package}
\title{dqshiny: Enhance shiny apps with customizable modules}
\description{
This package provides highly customizable modules to enhance
 your shiny apps. Add custom header, headline and footer to your app.
 Include collapsable boxes and value boxes independent from the chosen UI
 Layout.
}
\section{Section}{
 some more info
  TO DO document the package!!
}

