% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering.R
\name{get_filters}
\alias{get_filters}
\alias{text_filter}
\alias{range_filter}
\title{Get filter values from a given context}
\usage{
get_filters(input, context)

text_filter(df, values)

range_filter(df, ranges)
}
\arguments{
\item{input}{shiny input environment}

\item{context}{the context used to specify all ui elements}

\item{df}{data frame to filter}

\item{values}{character array with the filter values, should have length of
data or being named}

\item{ranges}{numeric (or numeric makeable) vector with the filter ranges,
should have length of data or being named}
}
\value{
getFilter: list of filter values

text_filter: filtered data frame

range_filter: filtered data frame
}
\description{
get_filters receives the values from all filters within the
given context and returns them in a list.

text_filter filters a given data frame with the given filter
values. Names of the given values vector should be the indices of the
corresponding data frame columns. All filters are case-ignoring.

range_filter filters a given data frame with the given filter
values. Names of the given ranges vector should be the indices of the
corresponding data frame columns.
}
