% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{init}
\alias{init}
\title{Initializes dqshiny styling}
\usage{
init()
}
\arguments{
\item{selectize}{optional logical to indicate whether selectize select and handsontable-selectize-editor should be loaded or not}
}
\value{
list of shiny includes executed by shinyUI
}
\description{
Must be used inside the shinyUI before any other daqana styling function is used.
Specify parameters to automatically add content to your side. Please notice, that no additional parameters can be given in this case.
}
\examples{
\donttest{library(shiny)
shinyApp(
  ui = fluidPage(
    init(),
    dq_header()
  ),
  server = function(input, output) {
  }
)}
}
\author{
richard.kunze
}
