% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{init_fonts}
\alias{init_fonts}
\title{Initializes dqshiny fonts for figures}
\usage{
init_fonts(paths, pattern = "ttf$", silent = TRUE)
}
\arguments{
\item{paths}{vector of paths to look in}

\item{pattern}{pattern to th files have to match}

\item{silent}{optional logical indicating whether message and warnings should be printed or not}
}
\description{
Only needed for highcharter/ggplot2 figures
}
\author{
david.breuer
}
