% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering.R
\name{range_filter}
\alias{range_filter}
\title{numeric range filter for data.frames}
\usage{
range_filter(df, ranges)
}
\arguments{
\item{df}{data frame to filter}

\item{ranges}{numeric (or convertable) vector with the filter ranges,
should have length of data or being named}
}
\value{
range_filter: filtered data frame
}
\description{
range_filter filters a given data frame with the given filter
values. Names of the given ranges vector should be the indices of the
corresponding data frame columns.
}
\author{
richard.kunze
}
