% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box_group.R
\name{render_dq_box_group}
\alias{render_dq_box_group}
\title{Directly render a set of dq_boxes as a group}
\usage{
render_dq_box_group(..., open = NULL)
}
\arguments{
\item{...}{a set of dq_boxes}

\item{open}{optional integer or character of length one, specifying the
initially opened box}
}
\value{
fluidRow containing the set of groupified dq_boxes
}
\description{
Directly render a set of dq_boxes into an uiOutput element. All
given dq_boxes will be groupified automatically, meaning that they will
become 'collapsible' and 'openCallback'ed.
}
\examples{
\donttest{library(shiny)
shinyApp(
  ui = fluidPage(
    init(),
    fluidRow(column(6, uiOutput("myGroup")))
  ),
  server = function(input, output) {
    output$myGroup <- render_dq_box_group(
      dq_box("Random Input1", dqSpace(250), "End of Content", title = "TestBox1", width = 12),
      dq_box("Random Input2", dqSpace(250), "End of Content", title = "TestBox2", width = 12),
      dq_box("Random Input3", dqSpace(250), "End of Content", title = "TestBox3", width = 12),
      open = 3L)
  }
)}
}
\author{
richard.kunze
}
