% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering.R
\name{text_filter}
\alias{text_filter}
\title{text filter for data.frames}
\usage{
text_filter(df, values, cols_filter = NULL, id_col_filter_placement = NULL)
}
\arguments{
\item{df}{data frame to filter}

\item{values}{character array with the filter values, should have length of
data or being named}

\item{cols_filter}{optional list of columns to be filtered simultaneously
using the first filter parameter with type "T"}

\item{id_col_filter_placement}{optional column index to place the simutaneous
filter, must be non-null when the parameter cols_filter is non-null}
}
\value{
text_filter: filtered data frame
}
\description{
text_filter filters a given data frame with the given filter
values. Names of the given values vector should be the indices of the
corresponding data frame columns. All filters are case-ignoring.
}
\author{
richard.kunze
}
