% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box_group.R
\name{update_dq_box}
\alias{update_dq_box}
\title{Function to update the collapsed status of a dqBox}
\usage{
update_dq_box(id, collapsed, silent = FALSE)
}
\arguments{
\item{id}{id of the dqBox}

\item{collapsed}{new collapsed status to be set}

\item{silent}{optional logical indicating to suppress events or not}
}
\value{
sent message
}
\description{
Function to send a message to js to update the collapsed status of a dqBox.
}
\author{
richard.kunze
}
