% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dqshiny.R
\docType{package}
\name{dqshiny}
\alias{dqshiny}
\alias{dqshiny-package}
\title{dqshiny: Enhance Shiny Apps with Customizable Modules}
\description{
Provides highly customizable modules to enhance your shiny apps.
Includes layout independent collapsible boxes and value boxes, a very fast
autocomplete input, rhandsontable extensions for filtering and paging and
much more.
}
\section{Section}{
 some more info
  TO DO document the package!!
}

