% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initital_state.R
\name{hidden}
\alias{hidden}
\alias{disabled}
\title{Sets the initial state of all given tags}
\usage{
hidden(...)

disabled(...)
}
\arguments{
\item{...}{tags to add to the ui, can be a single element or nested tagLists}
}
\value{
tags with the state change
}
\description{
Hidden only hides elements on the top level, so showing these elements will also show all children of them.

Disabled will recursively traverse the given elements and its children and set all inputs, buttons, selects and textfields
to be disabled.
}
\examples{
\donttest{library(shiny)
shinyApp(
  ui = fluidPage(
    init(),
    actionButton("btn1", "Press Me"),
    hidden(disabled(actionButton("btn2", "Show")))
  ),
  server = function(input, output) {
    observeEvent(input$btn1, toggle("btn2"))
  }
)}
}
\author{
richard.kunze
}
