% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{init}
\alias{init}
\title{Initializes dqshiny}
\usage{
init()
}
\value{
dqshiny CSS and JS dependency
}
\description{
Can be used inside the shinyUI to add dq shiny resource path.
Will be automatically called by loading the package and using any of the
dqshiny elements.
}
\examples{
## Only run examples in interactive R sessions
if (interactive()) {

library(shiny)
shinyApp(
  ui = fluidPage(
    init()
  ),
  server = function(input, output) {}
)

}
}
\author{
richard.kunze
}
